flask_restful_03_token_auth/requirements.txt
flask
flask-sqlalchemy
flask-restful
flask-httpauth
pyjwt
